Fix scripts/bacula.in to have awk on an environment variable   and add comments for Solaris users. - Turn off inet_aton in src/lib/address_conf.c for Win32 - Add new files to win32 build and eliminate a compiler warning. - Add sample DVD Device resource to bacula-sd.conf
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1799 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/bnet.c
@@ -936,7 +936,7 @@ const char *bnet_sig_to_ascii(BSOCK * bs)
    case BNET_PROMPT:
       return "BNET_PROMPT";
    default:
-      sprintf(buf, "Unknown sig %d", bs->msglen);
+      sprintf(buf, "Unknown sig %d", (int)bs->msglen);
       return buf;
    }
 }
